internal/infra-bionemo/requirements-test.txt
pytest-cov==5.0.0
